refactor: mv Tina => TinaProvider
